voice-sphinx-server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
numpy>=1.24.0
openai-whisper>=20231117
pydantic>=2.0.0
python-multipart>=0.0.6
torch>=2.0.0
soundfile>=0.12.1
loguru>=0.7.0
